+++ Assets/Tree/TreeGenerator.cs
@@ -1,9 +1,10 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using Tree;
+using TreeScripts;
+using UserScripts;
 
-namespace Tree {
+namespace TreeScripts {
   public class TreeGenerator : MonoBehaviour
   {
     [SerializeField]
@@ -11,35 +12,19 @@ namespace Tree {
     [SerializeField]
     GameObject leave;
     [SerializeField]
-    Color leaveBaseColor;
-    [SerializeField]
-    float baseColorRate = 0.5f;
-    [SerializeField]
     Transform treeParent;
     [SerializeField]
     int num, segmentNum, branchNum;
     [SerializeField]
     float radius, branchLength, stemRadius;
-    UserManager userManager;
-    void Start()
-    {
-      userManager = GetComponent<UserManager>();
-      StartCoroutine("MakeTree");
 
-      // 葉の色
-      for(int i=0;i<Points.GENRE_TO_COLOR.Length;i++) {
-        Points.GENRE_TO_COLOR[i] = Points.GENRE_TO_COLOR[i] * 0.95f + new Color(1f, 1f, 1f) * 0.05f;
-        Points.GENRE_TO_COLOR[i] = Points.GENRE_TO_COLOR[i] * baseColorRate + leaveBaseColor * (1f - baseColorRate);
-        // Points.GENRE_TO_COLOR[i] = Mathf.Pow(new Color(1f, 1f, 1f) - Points.GENRE_TO_COLOR[i], 2.0f);
-      }
-    }
-
-    IEnumerator MakeTree() {
+    public void MakeTree() {
       List<User> users = new List<User>();
       TreeMesh treeMesh = new TreeMesh(segmentNum, branchNum, branchLength, stemRadius);
       treeMesh.BuildMesh();
       if(treeParent.childCount == 0)
       {
+        // 自動配置
         float phi = 0f;
         for(int i=0;i<num;i++) {
           GameObject t = Instantiate(tree, treeParent);
@@ -59,27 +44,16 @@ namespace Tree {
                           Quaternion.AngleAxis(Random.Range(0f, 360f), new Vector3(0f, 1f, 0f));
           // Quaternion quat = Quaternion.AngleAxis(-theta / Mathf.PI / 2f * 360f, new Vector3(0f, 0f, 1f));
           t.transform.rotation = quat;
-          yield return new WaitForSeconds(.01f);
         }
       } else {
-        
+        // すでに配置されてるのをActiveに
         num = treeParent.childCount;
         for(int i=0;i<num;i++) {
           GameObject t = treeParent.GetChild(i).gameObject;
           users.Add(t.GetComponent<User>());
           treeMesh.SetMesh(t.GetComponent<ControlTree>(), leave);
-          
-          yield return new WaitForSeconds(.01f);
         }
       }
-
-      userManager.SetTrees(users);
-    }
-
-
-    void Update()
-    {
-      
     }
   }
