+++ Assets/Tree/TreeMesh.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace Tree {
+namespace TreeScripts {
 
   public class TreeMesh {
