+++ Assets/User.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d7b4815ce4a7ab14dbf2985f7afd0827
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
