+++ Assets/VisualizerSystem/SystemInitializer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 56834f765062cfd43804e663748aab3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
